refactor status check code
